Layed some groundwork for mouse window selection
Added new button for Select Window Mode on ProgramWindow.  Required adding a click event in the main window to account for the button click on the ProgramWindow.

Hide the Main Window if Select Window Mode is activated, also hides all ProgramWindow objects.

If Select Window Mode is activated when the Main Window is maximized, no ProgramWindows will be made visible.